Interpreter memory access safety checks are added; Interpreter receives an additional flag to signalize program interpretation failures; Keypad Test CHIP-8 program is added.

+++ emu-chip8-client/main.cpp
@@ -282,7 +282,7 @@ static RECT counterTextRect = {
 
 static void UpdateFrameRateCounter() {
 	ULONGLONG timeCurrent = GetTickCount64();
-	FLOAT timeDifference = timeCurrent - timeLastFrame;
+	FLOAT timeDifference = FLOAT(timeCurrent - timeLastFrame);
 
 	if (timeDifference >= 1000) {
 		TCHAR output[13] = { 0 };
